test: move tests to new directory
