+++ rust/Cargo.toml
@@ -13,7 +13,7 @@ categories = ["api-bindings", "web-programming"]
 authors = ["VITAP Students <sanjay@nullvijayawada.org>"]
 
 [lib]
-crate-type = ["cdylib", "staticlib"]
+crate-type = ["cdylib", "staticlib", "rlib"]
 
 [dependencies]
 flutter_rust_bridge = "=2.11.1"
