+++ rust/src/api/vtop/parser/course_page_parser.rs
@@ -1,5 +1,5 @@
-use scraper::{Html, Selector};
 use crate::api::vtop::types::course_page::*;
+use scraper::{Html, Selector};
 
 /// Parses the courses dropdown from the course page HTML response
 pub fn parse_courses_for_course_page(html: String) -> CoursesResponse {
@@ -7,13 +7,14 @@ pub fn parse_courses_for_course_page(html: String) -> CoursesResponse {
     let mut courses = Vec::new();
 
     // Select option elements within the courseCode select
-    let select_selector = Selector::parse("select#courseCode option, select[name='courseCode'] option")
-        .unwrap_or_else(|_| Selector::parse("option").unwrap());
-    
+    let select_selector =
+        Selector::parse("select#courseCode option, select[name='courseCode'] option")
+            .unwrap_or_else(|_| Selector::parse("option").unwrap());
+
     for option in document.select(&select_selector) {
         let value = option.value().attr("value").unwrap_or("").to_string();
         let label = option.text().collect::<String>().trim().to_string();
-        
+
         // Skip the placeholder option
         if value.is_empty() || label.starts_with("--") || label.starts_with("-- ") {
             continue;
@@ -25,8 +26,8 @@ pub fn parse_courses_for_course_page(html: String) -> CoursesResponse {
         let (course_code, course_title, course_type) = if parts.len() >= 3 {
             (
                 parts[0].trim().to_string(),
-                parts[1..parts.len()-1].join(" - ").trim().to_string(),
-                parts[parts.len()-1].trim().to_string(),
+                parts[1..parts.len() - 1].join(" - ").trim().to_string(),
+                parts[parts.len() - 1].trim().to_string(),
             )
         } else if parts.len() == 2 {
             (
@@ -59,11 +60,11 @@ pub fn parse_slots_for_course_page(html: String, semester_id: &str) -> SlotsResp
     // Parse slot options
     let slot_selector = Selector::parse("select#slotId option, select[name='slotId'] option")
         .unwrap_or_else(|_| Selector::parse("option").unwrap());
-    
+
     for option in document.select(&slot_selector) {
         let value = option.value().attr("value").unwrap_or("").to_string();
         let label = option.text().collect::<String>().trim().to_string();
-        
+
         // Skip the placeholder option
         if value.is_empty() || label.starts_with("--") || label.starts_with("-- ") {
             continue;
@@ -80,14 +81,19 @@ pub fn parse_slots_for_course_page(html: String, semester_id: &str) -> SlotsResp
 
     for table in document.select(&table_selector) {
         let rows: Vec<_> = table.select(&row_selector).collect();
-        
+
         // Skip header row
         for row in rows.iter().skip(1) {
             let cells: Vec<_> = row.select(&cell_selector).collect();
-            
+
             if cells.len() >= 9 {
                 // Extract data from cells
-                let sl_no = cells[0].text().collect::<String>().trim().parse::<i32>().unwrap_or(0);
+                let sl_no = cells[0]
+                    .text()
+                    .collect::<String>()
+                    .trim()
+                    .parse::<i32>()
+                    .unwrap_or(0);
                 let class_group = cells[1].text().collect::<String>().trim().to_string();
                 let course_code = cells[2].text().collect::<String>().trim().to_string();
                 let course_title = cells[3].text().collect::<String>().trim().to_string();
@@ -125,40 +131,47 @@ pub fn parse_slots_for_course_page(html: String, semester_id: &str) -> SlotsResp
         }
     }
 
-    SlotsResponse { slots, class_entries }
+    SlotsResponse {
+        slots,
+        class_entries,
+    }
 }
 
 /// Parses the detailed course page with all lectures and materials
 pub fn parse_course_detail_page(html: String) -> CoursePageDetail {
     let document = Html::parse_document(&html);
-    
+
     // Extract course info from the first table
     let course_info = extract_course_info(&document);
-    
+
     // Extract semester_id and class_id from hidden inputs
     let semester_id = extract_hidden_input_value(&document, "semesterSubId");
     let class_id = extract_hidden_input_value(&document, "classId");
     let course_id = extract_hidden_input_value(&document, "courseId");
-    
+
     // Update course_info with course_id
     let mut course_info = course_info;
     course_info.course_id = course_id;
-    
+
     // Extract download paths
     let download_all_path = extract_download_path(&document, "allCourseMeterialDownload/1/1");
-    let download_general_materials_path = extract_download_path(&document, "allCourseMeterialDownload/2/1");
+    let download_general_materials_path =
+        extract_download_path(&document, "allCourseMeterialDownload/2/1");
     let syllabus_download_path = extract_download_path(&document, "courseSyllabusDownload");
-    
+
     // Build course plan download path
     let course_plan_download_path = if !semester_id.is_empty() && !class_id.is_empty() {
-        Some(format!("academics/common/CoursePlanExcelDownload?semesterSubId={}&classId={}", semester_id, class_id))
+        Some(format!(
+            "academics/common/CoursePlanExcelDownload?semesterSubId={}&classId={}",
+            semester_id, class_id
+        ))
     } else {
         None
     };
-    
+
     // Extract lectures
     let lectures = extract_lectures(&document);
-    
+
     CoursePageDetail {
         course_info,
         semester_id,
@@ -190,11 +203,11 @@ fn extract_course_info(document: &Html) -> CourseInfo {
     // Find the first table with course info
     if let Some(table) = document.select(&table_selector).next() {
         let rows: Vec<_> = table.select(&row_selector).collect();
-        
+
         // Find the data row (skip header)
         for row in rows.iter().skip(1) {
             let cells: Vec<_> = row.select(&cell_selector).collect();
-            
+
             if cells.len() >= 7 {
                 course_info.class_group = cells[0].text().collect::<String>().trim().to_string();
                 course_info.course_code = cells[1].text().collect::<String>().trim().to_string();
@@ -213,9 +226,10 @@ fn extract_course_info(document: &Html) -> CourseInfo {
 
 /// Extracts hidden input value by name
 fn extract_hidden_input_value(document: &Html, name: &str) -> String {
-    let selector = Selector::parse(&format!("input[name='{}'], input#{}",name, name)).unwrap();
-    
-    document.select(&selector)
+    let selector = Selector::parse(&format!("input[name='{}'], input#{}", name, name)).unwrap();
+
+    document
+        .select(&selector)
         .next()
         .and_then(|input| input.value().attr("value"))
         .unwrap_or("")
@@ -224,8 +238,9 @@ fn extract_hidden_input_value(document: &Html, name: &str) -> String {
 
 /// Extracts download path from href containing the specified pattern
 fn extract_download_path(document: &Html, pattern: &str) -> Option<String> {
-    let link_selector = Selector::parse("a[href*='vtopDownload'], a[onclick*='vtopDownload']").unwrap();
-    
+    let link_selector =
+        Selector::parse("a[href*='vtopDownload'], a[onclick*='vtopDownload']").unwrap();
+
     for link in document.select(&link_selector) {
         // Check href attribute
         if let Some(href) = link.value().attr("href") {
@@ -236,7 +251,7 @@ fn extract_download_path(document: &Html, pattern: &str) -> Option<String> {
                 }
             }
         }
-        
+
         // Check onclick attribute
         if let Some(onclick) = link.value().attr("onclick") {
             if onclick.contains(pattern) {
@@ -246,7 +261,7 @@ fn extract_download_path(document: &Html, pattern: &str) -> Option<String> {
             }
         }
     }
-    
+
     None
 }
 
@@ -262,31 +277,38 @@ fn extract_lectures(document: &Html) -> Vec<LectureEntry> {
     // Find the table with lectures (it has columns: Sl.No., Lecture Date, Lecture Day, Lecture Topic, Reference Material)
     for table in document.select(&table_selector) {
         let rows: Vec<_> = table.select(&row_selector).collect();
-        
+
         // Check if this is the lecture table by examining headers
         if rows.is_empty() {
             continue;
         }
-        
+
         let header_text = rows[0].text().collect::<String>();
         if !header_text.contains("Lecture Date") && !header_text.contains("Lecture Topic") {
             continue;
         }
-        
+
         // Process lecture rows
         for row in rows.iter().skip(1) {
             let cells: Vec<_> = row.select(&cell_selector).collect();
-            
+
             if cells.len() >= 5 {
-                let sl_no = cells[0].text().collect::<String>().trim().parse::<i32>().unwrap_or(0);
-                
+                let sl_no = cells[0]
+                    .text()
+                    .collect::<String>()
+                    .trim()
+                    .parse::<i32>()
+                    .unwrap_or(0);
+
                 // Extract date - there are two spans inside
                 let date;
                 let formatted_date;
                 let spans: Vec<_> = cells[1].select(&span_selector).collect();
                 if spans.len() >= 2 {
                     date = spans[0].text().collect::<String>().trim().to_string();
-                    formatted_date = spans[1].text().collect::<String>()
+                    formatted_date = spans[1]
+                        .text()
+                        .collect::<String>()
                         .trim()
                         .trim_start_matches('[')
                         .trim_end_matches(']')
@@ -298,10 +320,10 @@ fn extract_lectures(document: &Html) -> Vec<LectureEntry> {
                     date = cells[1].text().collect::<String>().trim().to_string();
                     formatted_date = date.clone();
                 }
-                
+
                 let day = cells[2].text().collect::<String>().trim().to_string();
                 let topic = cells[3].text().collect::<String>().trim().to_string();
-                
+
                 // Extract reference materials
                 let mut reference_materials = Vec::new();
                 for link in cells[4].select(&link_selector) {
@@ -317,7 +339,7 @@ fn extract_lectures(document: &Html) -> Vec<LectureEntry> {
                         }
                     }
                 }
-                
+
                 if sl_no > 0 {
                     lectures.push(LectureEntry {
                         sl_no,
@@ -330,13 +352,13 @@ fn extract_lectures(document: &Html) -> Vec<LectureEntry> {
                 }
             }
         }
-        
+
         // Found the lecture table, no need to continue
         if !lectures.is_empty() {
             break;
         }
     }
-    
+
     lectures
 }
 
@@ -347,12 +369,20 @@ fn parse_erp_id_from_onclick(onclick: &str) -> Option<String> {
         // Extract content between parentheses
         let start = onclick.find('(')?;
         let end = onclick.rfind(')')?;
-        let params = &onclick[start+1..end];
-        
+        let params = &onclick[start + 1..end];
+
         // Split by comma and extract the second parameter (erp_id)
         let parts: Vec<&str> = params.split(',').collect();
         if parts.len() >= 2 {
-            return Some(parts[1].trim().trim_matches('\'').trim_matches('&').trim_matches('#').trim_matches(';').to_string());
+            return Some(
+                parts[1]
+                    .trim()
+                    .trim_matches('\'')
+                    .trim_matches('&')
+                    .trim_matches('#')
+                    .trim_matches(';')
+                    .to_string(),
+            );
         }
     }
     None
@@ -362,17 +392,16 @@ fn parse_erp_id_from_onclick(onclick: &str) -> Option<String> {
 fn extract_path_from_vtop_download(href: &str) -> Option<String> {
     // Format: javascript:vtopDownload('path/here')
     // or: vtopDownload('path/here')
-    
+
     // Handle HTML entity encoding &#39; for single quotes
     let cleaned = href.replace("&#39;", "'");
-    
+
     let start = cleaned.find("vtopDownload(")?;
     let path_start = cleaned[start..].find('\'')? + start + 1;
     let path_end = cleaned[path_start..].find('\'')? + path_start;
-    
+
     Some(cleaned[path_start..path_end].to_string())
 }
-
 #[cfg(test)]
 mod tests {
     use super::*;
